Importer des données depuis le presse-papier

Fonction qui permet d'importer sous R des données depuis le presse-papiers. Ces données peuvent par exemple venir d'un tableur (Excel, Calc, etc.) : copier les données du tableur et ensuite lancer sous R la fonction ctr_v()

Utilisation

ctrl_v(header = TRUE, row.names, stringsAsFactors = FALSE, dec = ",", ...)

Arguments

header

booléen indiquant si la première ligne doit être interprétée comme les noms des colonnes (TRUE par défaut). Paramètre utilisé uniquement si les données du presse-papier contiennent plusieurs colonnes.

row.names

un vecteur contenant les noms des lignes. Cela peut être un vecteur contenant les noms des lignes, un entier donnant le numéro de la colonne contenant les noms des lignes ou une chaîne de caractères donnant le nom de la colonne qui contient les noms des lignes. Lorsque row.names = NULL alors les lignes sont numérotées.

stringsAsFactors

booléen indiquant si l'on souhaite que les chaînes de charactères soient transformées en facteur dans la table de sortie (FALSE par défaut). Paramètre utilisé uniquement si les données du presse-papier contiennent plusieurs colonnes.

dec

le caractère utilisé comme séparateur décimal dans les données du presse-papiers. Par défaut la virgule est utilisée : utiliser le point pour si les données proviennent d'un tableur non francophone. Paramètre utilisé uniquement si les données du presse-papier contiennent plusieurs colonnes.

...

autres paramètres de read.table (uniquement si les données du presse-papier contiennent plusieurs colonnes).

Valeur de retour

Un vecteur de caractères s'il n'y a qu'une colonne, sinon un data.frame.

Voir également

Exemples

# Après avec copié des données depuis un tableur
ctrl_v()
#> Error in ctrl_v(): objet 'tableau' introuvable