Graphiques interactifs d'une série temporelle avec highcharter

Fonctions qui permettent de faire des graphiques interactifs de séries temporelles à partir d'un objet de type ts.

Utilisation

hc_lines(
  data,
  titre = NULL,
  sous_titre = NULL,
  legende = NULL,
  affiche_legende = TRUE,
  x_lab = NULL,
  y_lab = NULL,
  outDec = ",",
  useHTML = FALSE,
  digits = 1
)

hc_stocks(
  data,
  titre = NULL,
  sous_titre = NULL,
  legende = NULL,
  affiche_legende = TRUE,
  x_lab = NULL,
  y_lab = NULL,
  outDec = ",",
  useHTML = FALSE,
  type = NULL,
  color = NULL,
  digits = 1
)

Arguments

data

un objet de type ts contenant une ou plusieurs séries à grapher.

titre

titre du graphique (par défaut pas de titre).

sous_titre

sous-titre du graphique (par défaut pas de sous-titre).

legende

légende du graphique. Par défaut le nom des variables dans l'objet data en entrée. S'il n'y a qu'une série à tracer alors aucune légende n'est affichée.

affiche_legende

booléen indiquant si l'on affiche la légende.

x_lab

titre axe des abscisses (par défaut aucun titre).

y_lab

titre axe des ordonnées (par défaut "Date").

outDec

séparateur décimal utilisé pour dans la légende des axes (par défaut la virgule).

useHTML

Booléen indiquant si l'on veut utiliser du code HTML dans les titres.

digits

nombre de chiffres après la virgule.

type

vecteur contenant le type à utiliser pour les séries (par défaut "lines").

color

vecteur contenant les couleurs à utiliser (par défaut on garde les couleurs de highchart).

Valeur de retour

Un graphique highcharter.

Exemples

data <- lectureBDM("001585942","001585980")
legende <- c("Carnets de commandes globaux","Carnets de commandes étrangers")
titre <- "Soldes d'opinion sur les carnets de commandes dans l'industrie manufacturière"
hc_lines(data, titre = titre, legende = legende)

hc_stocks(data, titre = titre, legende = legende)