Convertisseur tableau en ts

Fonctions qui permettent de convertir des tableaux en ts.

Utilisation

ymd_ts(x, col_date = 1, sep_date, frequence = NULL)

dmy_ts(x, col_date = 1, sep_date, frequence = NULL)

Arguments

x

table contenant au moins deux colonnes dont une avec des dates.

col_date

numéro ou nom de la colonne contenant les dates (par défaut c'est la première colonne).

sep_date

caractère utilisé pour séparé le jour, le mois et l'année. Détection automatique par défaut.

frequence

fréquence de l'objet ts. Par défaut la fonction fait une recherche automatique (il peut y avoir des problèmes si l'on n'a pas d'année entière connue).

Valeur de retour

Un ts.

Détails

Les fonctions ymd_ts() et dmy_ts() permettent de convertir des tableaux (par exemple des data.frame) en ts. Pour cela il faut dans le tableau d'entrée une colonne contenant les dates qui est au format YYYY-MM-JJ (fonction ymd_ts()) ou au format JJ-MM-YYYY (fonction dmy_ts()). Le séparateur utilisé entre le jour, le mois et l'année peut être modifié par le paramètre sep_date.

Exemples

x <- data.frame(dates=paste(rep(2000:2002, each=4),
                            rep(c("01", "04", "07", "10"),2) ,"01", sep = "-"),
                donnees = 1:12, stringsAsFactors = FALSE)
ymd_ts(x)
#>      Qtr1 Qtr2 Qtr3 Qtr4
#> 2000    1    2    3    4
#> 2001    5    6    7    8
#> 2002    9   10   11   12