Téléchargement d'une série de la BDM Insee

Fonction qui permet d'importer des séries de la Banque de données macro-économiques (BDM) de l'Insee à partir de son idbank. L'idbank d'une série peut être trouvé sur le site de la bdm (https://www.bdm.insee.fr/bdm2/index). permet également l'utilisation de régressions quantiles.

Utilisation

lectureBDM(idbank, ...)

Arguments

idbank

un vecteur d'un ou plusieurs idbank à télécharger.

...

d'autres vecteurs d'un ou plusieurs idbank à télécharger.

Valeur de retour

Retourne un objet de type ts ou mts, en fonction du nombre d'idbank en entrée, de la même fréquence que les séries en entrée.

Détails

La fonction permet de télécharger des séries temporelles présentes la Banque de données macro-économiques (BDM) de l'Insee à partir de son idbank. Elle utilise pour cela le service web SDMX de l'Insee (voir http://www.bdm.insee.fr/bdm2/statique?page=sdmx pour plus d'informations). Elle utilise le package rsdmx pour la lecture des données et reshape2 pour la mise en forme. Les idbank en paramètre doivent correspondre à des séries qui ont la même périodicité. Si un des idbank en entrée n'est pas trouvé alors la fonction renvoit également un message d'avertissement.

Exemples

# Pour télécharger l'IPI manufacturier CVS : 001654241
lectureBDM("001654241")
#>         Jan    Feb    Mar    Apr    May    Jun    Jul    Aug    Sep    Oct
#> 1990 117.40 119.51 118.29 118.35 118.81 117.57 116.70 115.76 116.74 116.26
#> 1991 116.94 115.48 116.02 114.65 112.73 114.57 113.18 114.45 114.86 113.21
#> 1992 112.87 112.67 113.67 112.29 113.31 111.42 110.29 111.08 108.87 109.05
#> 1993 107.35 106.34 105.35 105.03 104.63 103.86 106.15 104.29 105.43 104.78
#> 1994 105.66 105.15 105.10 107.53 108.25 107.82 108.40 108.67 107.99 109.77
#> 1995 110.94 111.56 110.54 111.45 110.16 110.61 112.08 110.07 111.33 110.22
#> 1996 108.55 107.40 109.17 107.66 109.24 109.02 106.42 109.67 108.66 107.59
#> 1997 107.21 109.31 110.30 112.95 111.37 111.76 111.20 115.93 113.06 115.45
#> 1998 115.50 115.74 113.98 114.87 116.36 115.75 116.03 113.98 115.21 114.50
#> 1999 114.76 113.00 113.33 114.19 114.39 115.66 115.58 113.72 116.13 117.93
#> 2000 116.81 117.16 117.56 118.32 119.85 118.07 119.19 117.13 118.70 120.00
#> 2001 120.09 120.48 120.77 117.81 118.89 119.02 116.95 119.91 117.60 117.14
#> 2002 116.03 116.52 116.73 117.13 115.85 115.15 114.46 116.86 114.54 113.29
#> 2003 113.78 114.20 114.32 113.91 110.96 110.75 113.00 112.00 112.00 113.93
#> 2004 113.19 115.13 114.01 113.99 113.02 114.69 115.46 111.58 114.66 115.49
#> 2005 116.00 113.64 111.29 114.63 112.99 112.89 112.97 112.41 115.08 111.91
#> 2006 113.71 113.10 115.55 113.82 117.00 117.12 115.01 115.52 116.68 115.67
#> 2007 115.83 116.74 117.56 116.15 118.53 117.65 117.71 118.95 115.19 117.69
#> 2008 117.36 118.75 116.77 118.20 113.57 113.04 113.81 111.82 111.50 109.09
#> 2009  96.23  95.10  94.12  93.85  95.17  96.82  96.00  96.77  98.34  97.26
#> 2010  97.69  97.27  99.59  99.62 100.17 100.05 100.96  99.02 101.04 100.56
#> 2011 104.80 105.35 104.31 102.98 106.49 102.32 103.77 102.45 102.37 102.38
#> 2012 102.22 100.87 102.69 101.18 100.86 100.73 101.63 103.32 100.75  99.27
#> 2013  99.42  99.60  99.14 101.32 101.29 100.33 100.71  99.36 100.03 100.25
#> 2014  99.66 100.64 100.03 100.58  97.92  99.66 100.96  98.72 100.90  99.56
#> 2015 100.27 100.38 101.35 100.89 101.32 102.27 100.37 102.97 102.99 102.70
#> 2016 103.21 102.00 100.50 102.66 102.01 101.07 100.69 103.40 101.91 100.84
#> 2017 102.28 101.44 104.40 103.01 105.12 104.18 104.79 104.59 105.33 107.95
#>         Nov    Dec
#> 1990 114.41 115.80
#> 1991 115.26 114.26
#> 1992 108.63 106.80
#> 1993 103.19 102.70
#> 1994 111.09 110.40
#> 1995 108.27 111.16
#> 1996 108.89 108.78
#> 1997 113.34 115.07
#> 1998 114.65 113.66
#> 1999 117.60 116.34
#> 2000 119.34 121.43
#> 2001 114.97 115.18
#> 2002 114.81 112.02
#> 2003 111.89 112.61
#> 2004 113.25 113.45
#> 2005 116.31 114.87
#> 2006 114.82 117.07
#> 2007 116.33 116.30
#> 2008 102.76 101.00
#> 2009  98.79  97.38
#> 2010 101.45 102.58
#> 2011 104.84 103.58
#> 2012  99.11 100.10
#> 2013 100.59 100.21
#> 2014  98.36 101.59
#> 2015 101.95 102.54
#> 2016 103.90 102.61
#> 2017 107.11 107.44
# Pour télécharger les IPI dans l'industrie manufacturière et dans l'agroalimentaire :
lectureBDM(c("001654241","001654289"))
#>          001654241 001654289
#> Jan 1990    117.40     90.76
#> Feb 1990    119.51     94.44
#> Mar 1990    118.29     94.36
#> Apr 1990    118.35     95.43
#> May 1990    118.81     97.30
#> Jun 1990    117.57     93.91
#> Jul 1990    116.70     94.66
#> Aug 1990    115.76     95.76
#> Sep 1990    116.74     92.77
#> Oct 1990    116.26     93.54
#> Nov 1990    114.41     92.89
#> Dec 1990    115.80     93.26
#> Jan 1991    116.94     98.25
#> Feb 1991    115.48     97.01
#> Mar 1991    116.02     95.74
#> Apr 1991    114.65     97.27
#> May 1991    112.73     94.45
#> Jun 1991    114.57     93.38
#> Jul 1991    113.18     94.55
#> Aug 1991    114.45     94.85
#> Sep 1991    114.86     96.04
#> Oct 1991    113.21     93.81
#> Nov 1991    115.26     95.30
#> Dec 1991    114.26     94.50
#> Jan 1992    112.87     95.52
#> Feb 1992    112.67     95.07
#> Mar 1992    113.67     95.49
#> Apr 1992    112.29     96.11
#> May 1992    113.31     97.80
#> Jun 1992    111.42     96.55
#> Jul 1992    110.29     95.41
#> Aug 1992    111.08     93.32
#> Sep 1992    108.87     93.35
#> Oct 1992    109.05     93.05
#> Nov 1992    108.63     93.61
#> Dec 1992    106.80     93.94
#> Jan 1993    107.35     91.64
#> Feb 1993    106.34     92.82
#> Mar 1993    105.35     92.47
#> Apr 1993    105.03     92.39
#> May 1993    104.63     89.88
#> Jun 1993    103.86     91.85
#> Jul 1993    106.15     92.50
#> Aug 1993    104.29     92.84
#> Sep 1993    105.43     95.14
#> Oct 1993    104.78     93.15
#> Nov 1993    103.19     92.21
#> Dec 1993    102.70     91.19
#> Jan 1994    105.66     92.75
#> Feb 1994    105.15     91.66
#> Mar 1994    105.10     92.24
#> Apr 1994    107.53     91.42
#> May 1994    108.25     91.53
#> Jun 1994    107.82     92.34
#> Jul 1994    108.40     95.66
#> Aug 1994    108.67     98.13
#> Sep 1994    107.99     93.23
#> Oct 1994    109.77     93.83
#> Nov 1994    111.09     95.94
#> Dec 1994    110.40     94.76
#> Jan 1995    110.94     94.51
#> Feb 1995    111.56     94.83
#> Mar 1995    110.54     93.63
#> Apr 1995    111.45     95.25
#> May 1995    110.16     95.92
#> Jun 1995    110.61     96.47
#> Jul 1995    112.08     95.63
#> Aug 1995    110.07     96.13
#> Sep 1995    111.33     95.98
#> Oct 1995    110.22     96.45
#> Nov 1995    108.27     94.67
#> Dec 1995    111.16     95.24
#> Jan 1996    108.55     96.56
#> Feb 1996    107.40     97.13
#> Mar 1996    109.17     96.63
#> Apr 1996    107.66     96.47
#> May 1996    109.24    100.64
#> Jun 1996    109.02     96.75
#> Jul 1996    106.42     96.34
#> Aug 1996    109.67     98.13
#> Sep 1996    108.66     98.17
#> Oct 1996    107.59     97.30
#> Nov 1996    108.89     96.94
#> Dec 1996    108.78    100.23
#> Jan 1997    107.21     96.91
#> Feb 1997    109.31     97.95
#> Mar 1997    110.30     97.92
#> Apr 1997    112.95    100.22
#> May 1997    111.37     98.96
#> Jun 1997    111.76     98.42
#> Jul 1997    111.20     96.34
#> Aug 1997    115.93     99.65
#> Sep 1997    113.06     99.76
#> Oct 1997    115.45    101.31
#> Nov 1997    113.34     99.47
#> Dec 1997    115.07    102.33
#> Jan 1998    115.50    101.77
#> Feb 1998    115.74    100.27
#> Mar 1998    113.98    100.22
#> Apr 1998    114.87     99.66
#> May 1998    116.36     99.12
#> Jun 1998    115.75    100.16
#> Jul 1998    116.03    100.80
#> Aug 1998    113.98     98.17
#> Sep 1998    115.21     97.92
#> Oct 1998    114.50     99.07
#> Nov 1998    114.65     97.69
#> Dec 1998    113.66     97.96
#> Jan 1999    114.76     99.80
#> Feb 1999    113.00     99.25
#> Mar 1999    113.33     99.95
#> Apr 1999    114.19     97.91
#> May 1999    114.39     97.90
#> Jun 1999    115.66     99.88
#> Jul 1999    115.58    100.88
#> Aug 1999    113.72     99.03
#> Sep 1999    116.13    101.41
#> Oct 1999    117.93    103.12
#> Nov 1999    117.60    102.95
#> Dec 1999    116.34     95.78
#> Jan 2000    116.81     96.56
#> Feb 2000    117.16     97.58
#> Mar 2000    117.56     97.82
#> Apr 2000    118.32     98.25
#> May 2000    119.85     98.53
#> Jun 2000    118.07     98.74
#> Jul 2000    119.19    100.25
#> Aug 2000    117.13    100.25
#> Sep 2000    118.70     99.51
#> Oct 2000    120.00     99.35
#> Nov 2000    119.34     96.65
#> Dec 2000    121.43     99.16
#> Jan 2001    120.09     96.38
#> Feb 2001    120.48     97.90
#> Mar 2001    120.77     97.07
#> Apr 2001    117.81    100.71
#> May 2001    118.89    101.28
#> Jun 2001    119.02    100.32
#> Jul 2001    116.95     99.40
#> Aug 2001    119.91    101.19
#> Sep 2001    117.60     99.76
#> Oct 2001    117.14    100.99
#> Nov 2001    114.97    100.01
#> Dec 2001    115.18    101.74
#> Jan 2002    116.03    101.35
#> Feb 2002    116.52    101.66
#> Mar 2002    116.73    103.03
#> Apr 2002    117.13    102.03
#> May 2002    115.85    102.45
#> Jun 2002    115.15     99.10
#> Jul 2002    114.46     99.91
#> Aug 2002    116.86    102.49
#> Sep 2002    114.54    100.12
#> Oct 2002    113.29    101.34
#> Nov 2002    114.81    101.06
#> Dec 2002    112.02    100.55
#> Jan 2003    113.78    100.39
#> Feb 2003    114.20    101.44
#> Mar 2003    114.32    100.52
#> Apr 2003    113.91    101.79
#> May 2003    110.96    100.21
#> Jun 2003    110.75    100.87
#> Jul 2003    113.00    101.11
#> Aug 2003    112.00     98.55
#> Sep 2003    112.00    101.62
#> Oct 2003    113.93     99.70
#> Nov 2003    111.89     99.07
#> Dec 2003    112.61    100.22
#> Jan 2004    113.19    102.07
#> Feb 2004    115.13    100.30
#> Mar 2004    114.01    100.46
#> Apr 2004    113.99     99.58
#> May 2004    113.02     97.79
#> Jun 2004    114.69    100.45
#> Jul 2004    115.46    100.47
#> Aug 2004    111.58     98.05
#> Sep 2004    114.66     99.52
#> Oct 2004    115.49     99.17
#> Nov 2004    113.25    100.14
#> Dec 2004    113.45     98.41
#> Jan 2005    116.00     99.03
#> Feb 2005    113.64     98.21
#> Mar 2005    111.29     97.29
#> Apr 2005    114.63     97.85
#> May 2005    112.99     98.20
#> Jun 2005    112.89     97.69
#> Jul 2005    112.97     97.63
#> Aug 2005    112.41     99.65
#> Sep 2005    115.08     99.04
#> Oct 2005    111.91     98.29
#> Nov 2005    116.31    100.70
#> Dec 2005    114.87     99.18
#> Jan 2006    113.71     96.70
#> Feb 2006    113.10     97.68
#> Mar 2006    115.55     99.41
#> Apr 2006    113.82     98.55
#> May 2006    117.00     99.42
#> Jun 2006    117.12     98.44
#> Jul 2006    115.01    100.23
#> Aug 2006    115.52    100.48
#> Sep 2006    116.68     99.13
#> Oct 2006    115.67     99.56
#> Nov 2006    114.82     99.19
#> Dec 2006    117.07    100.05
#> Jan 2007    115.83    100.42
#> Feb 2007    116.74    100.87
#> Mar 2007    117.56    100.47
#> Apr 2007    116.15    102.49
#> May 2007    118.53    102.32
#> Jun 2007    117.65    101.16
#> Jul 2007    117.71     99.33
#> Aug 2007    118.95    102.21
#> Sep 2007    115.19     99.20
#> Oct 2007    117.69    102.73
#> Nov 2007    116.33    100.47
#> Dec 2007    116.30    101.62
#> Jan 2008    117.36    101.19
#> Feb 2008    118.75    100.96
#> Mar 2008    116.77    100.15
#> Apr 2008    118.20    100.62
#> May 2008    113.57     98.60
#> Jun 2008    113.04     97.14
#> Jul 2008    113.81     99.51
#> Aug 2008    111.82     98.29
#> Sep 2008    111.50     99.46
#> Oct 2008    109.09     99.21
#> Nov 2008    102.76     96.48
#> Dec 2008    101.00    100.86
#> Jan 2009     96.23     96.79
#> Feb 2009     95.10     97.72
#> Mar 2009     94.12     97.99
#> Apr 2009     93.85     97.51
#> May 2009     95.17     99.58
#> Jun 2009     96.82     99.51
#> Jul 2009     96.00     99.37
#> Aug 2009     96.77     97.86
#> Sep 2009     98.34     99.27
#> Oct 2009     97.26     98.02
#> Nov 2009     98.79     99.44
#> Dec 2009     97.38     97.60
#> Jan 2010     97.69     99.35
#> Feb 2010     97.27     99.27
#> Mar 2010     99.59    100.16
#> Apr 2010     99.62     98.94
#> May 2010    100.17    100.71
#> Jun 2010    100.05     99.45
#> Jul 2010    100.96    101.76
#> Aug 2010     99.02     99.46
#> Sep 2010    101.04    100.69
#> Oct 2010    100.56    100.43
#> Nov 2010    101.45    100.99
#> Dec 2010    102.58     98.79
#> Jan 2011    104.80    103.10
#> Feb 2011    105.35    102.64
#> Mar 2011    104.31    103.81
#> Apr 2011    102.98    102.12
#> May 2011    106.49    102.82
#> Jun 2011    102.32    104.54
#> Jul 2011    103.77    105.44
#> Aug 2011    102.45    102.10
#> Sep 2011    102.37    103.68
#> Oct 2011    102.38    104.07
#> Nov 2011    104.84    104.23
#> Dec 2011    103.58    101.81
#> Jan 2012    102.22    101.43
#> Feb 2012    100.87    102.89
#> Mar 2012    102.69    103.33
#> Apr 2012    101.18    103.21
#> May 2012    100.86    103.20
#> Jun 2012    100.73    103.31
#> Jul 2012    101.63    103.15
#> Aug 2012    103.32    103.49
#> Sep 2012    100.75    102.67
#> Oct 2012     99.27    104.24
#> Nov 2012     99.11    102.87
#> Dec 2012    100.10    104.88
#> Jan 2013     99.42    100.99
#> Feb 2013     99.60    101.53
#> Mar 2013     99.14    100.92
#> Apr 2013    101.32    102.00
#> May 2013    101.29    102.14
#> Jun 2013    100.33     96.73
#> Jul 2013    100.71    100.19
#> Aug 2013     99.36    101.25
#> Sep 2013    100.03    100.05
#> Oct 2013    100.25    100.38
#> Nov 2013    100.59    101.23
#> Dec 2013    100.21    101.79
#> Jan 2014     99.66    101.08
#> Feb 2014    100.64    102.10
#> Mar 2014    100.03    100.15
#> Apr 2014    100.58    102.62
#> May 2014     97.92    100.64
#> Jun 2014     99.66    100.69
#> Jul 2014    100.96    101.38
#> Aug 2014     98.72    100.20
#> Sep 2014    100.90    102.45
#> Oct 2014     99.56    100.74
#> Nov 2014     98.36     99.80
#> Dec 2014    101.59    101.80
#> Jan 2015    100.27    101.23
#> Feb 2015    100.38    100.05
#> Mar 2015    101.35    102.50
#> Apr 2015    100.89    101.99
#> May 2015    101.32    101.37
#> Jun 2015    102.27    103.65
#> Jul 2015    100.37    101.67
#> Aug 2015    102.97    102.29
#> Sep 2015    102.99    102.34
#> Oct 2015    102.70    102.32
#> Nov 2015    101.95    102.05
#> Dec 2015    102.54    101.20
#> Jan 2016    103.21    102.16
#> Feb 2016    102.00    102.46
#> Mar 2016    100.50    100.55
#> Apr 2016    102.66    100.85
#> May 2016    102.01     99.29
#> Jun 2016    101.07     99.02
#> Jul 2016    100.69     99.98
#> Aug 2016    103.40    101.02
#> Sep 2016    101.91    101.66
#> Oct 2016    100.84    100.03
#> Nov 2016    103.90    101.18
#> Dec 2016    102.61    101.07
#> Jan 2017    102.28     99.69
#> Feb 2017    101.44    100.72
#> Mar 2017    104.40    101.80
#> Apr 2017    103.01    100.48
#> May 2017    105.12    102.07
#> Jun 2017    104.18    101.03
#> Jul 2017    104.79    100.61
#> Aug 2017    104.59    101.25
#> Sep 2017    105.33    100.78
#> Oct 2017    107.95    101.05
#> Nov 2017    107.11    100.39
#> Dec 2017    107.44     99.15
# ou bien :
lectureBDM("001654241","001654289")
#>          001654241 001654289
#> Jan 1990    117.40     90.76
#> Feb 1990    119.51     94.44
#> Mar 1990    118.29     94.36
#> Apr 1990    118.35     95.43
#> May 1990    118.81     97.30
#> Jun 1990    117.57     93.91
#> Jul 1990    116.70     94.66
#> Aug 1990    115.76     95.76
#> Sep 1990    116.74     92.77
#> Oct 1990    116.26     93.54
#> Nov 1990    114.41     92.89
#> Dec 1990    115.80     93.26
#> Jan 1991    116.94     98.25
#> Feb 1991    115.48     97.01
#> Mar 1991    116.02     95.74
#> Apr 1991    114.65     97.27
#> May 1991    112.73     94.45
#> Jun 1991    114.57     93.38
#> Jul 1991    113.18     94.55
#> Aug 1991    114.45     94.85
#> Sep 1991    114.86     96.04
#> Oct 1991    113.21     93.81
#> Nov 1991    115.26     95.30
#> Dec 1991    114.26     94.50
#> Jan 1992    112.87     95.52
#> Feb 1992    112.67     95.07
#> Mar 1992    113.67     95.49
#> Apr 1992    112.29     96.11
#> May 1992    113.31     97.80
#> Jun 1992    111.42     96.55
#> Jul 1992    110.29     95.41
#> Aug 1992    111.08     93.32
#> Sep 1992    108.87     93.35
#> Oct 1992    109.05     93.05
#> Nov 1992    108.63     93.61
#> Dec 1992    106.80     93.94
#> Jan 1993    107.35     91.64
#> Feb 1993    106.34     92.82
#> Mar 1993    105.35     92.47
#> Apr 1993    105.03     92.39
#> May 1993    104.63     89.88
#> Jun 1993    103.86     91.85
#> Jul 1993    106.15     92.50
#> Aug 1993    104.29     92.84
#> Sep 1993    105.43     95.14
#> Oct 1993    104.78     93.15
#> Nov 1993    103.19     92.21
#> Dec 1993    102.70     91.19
#> Jan 1994    105.66     92.75
#> Feb 1994    105.15     91.66
#> Mar 1994    105.10     92.24
#> Apr 1994    107.53     91.42
#> May 1994    108.25     91.53
#> Jun 1994    107.82     92.34
#> Jul 1994    108.40     95.66
#> Aug 1994    108.67     98.13
#> Sep 1994    107.99     93.23
#> Oct 1994    109.77     93.83
#> Nov 1994    111.09     95.94
#> Dec 1994    110.40     94.76
#> Jan 1995    110.94     94.51
#> Feb 1995    111.56     94.83
#> Mar 1995    110.54     93.63
#> Apr 1995    111.45     95.25
#> May 1995    110.16     95.92
#> Jun 1995    110.61     96.47
#> Jul 1995    112.08     95.63
#> Aug 1995    110.07     96.13
#> Sep 1995    111.33     95.98
#> Oct 1995    110.22     96.45
#> Nov 1995    108.27     94.67
#> Dec 1995    111.16     95.24
#> Jan 1996    108.55     96.56
#> Feb 1996    107.40     97.13
#> Mar 1996    109.17     96.63
#> Apr 1996    107.66     96.47
#> May 1996    109.24    100.64
#> Jun 1996    109.02     96.75
#> Jul 1996    106.42     96.34
#> Aug 1996    109.67     98.13
#> Sep 1996    108.66     98.17
#> Oct 1996    107.59     97.30
#> Nov 1996    108.89     96.94
#> Dec 1996    108.78    100.23
#> Jan 1997    107.21     96.91
#> Feb 1997    109.31     97.95
#> Mar 1997    110.30     97.92
#> Apr 1997    112.95    100.22
#> May 1997    111.37     98.96
#> Jun 1997    111.76     98.42
#> Jul 1997    111.20     96.34
#> Aug 1997    115.93     99.65
#> Sep 1997    113.06     99.76
#> Oct 1997    115.45    101.31
#> Nov 1997    113.34     99.47
#> Dec 1997    115.07    102.33
#> Jan 1998    115.50    101.77
#> Feb 1998    115.74    100.27
#> Mar 1998    113.98    100.22
#> Apr 1998    114.87     99.66
#> May 1998    116.36     99.12
#> Jun 1998    115.75    100.16
#> Jul 1998    116.03    100.80
#> Aug 1998    113.98     98.17
#> Sep 1998    115.21     97.92
#> Oct 1998    114.50     99.07
#> Nov 1998    114.65     97.69
#> Dec 1998    113.66     97.96
#> Jan 1999    114.76     99.80
#> Feb 1999    113.00     99.25
#> Mar 1999    113.33     99.95
#> Apr 1999    114.19     97.91
#> May 1999    114.39     97.90
#> Jun 1999    115.66     99.88
#> Jul 1999    115.58    100.88
#> Aug 1999    113.72     99.03
#> Sep 1999    116.13    101.41
#> Oct 1999    117.93    103.12
#> Nov 1999    117.60    102.95
#> Dec 1999    116.34     95.78
#> Jan 2000    116.81     96.56
#> Feb 2000    117.16     97.58
#> Mar 2000    117.56     97.82
#> Apr 2000    118.32     98.25
#> May 2000    119.85     98.53
#> Jun 2000    118.07     98.74
#> Jul 2000    119.19    100.25
#> Aug 2000    117.13    100.25
#> Sep 2000    118.70     99.51
#> Oct 2000    120.00     99.35
#> Nov 2000    119.34     96.65
#> Dec 2000    121.43     99.16
#> Jan 2001    120.09     96.38
#> Feb 2001    120.48     97.90
#> Mar 2001    120.77     97.07
#> Apr 2001    117.81    100.71
#> May 2001    118.89    101.28
#> Jun 2001    119.02    100.32
#> Jul 2001    116.95     99.40
#> Aug 2001    119.91    101.19
#> Sep 2001    117.60     99.76
#> Oct 2001    117.14    100.99
#> Nov 2001    114.97    100.01
#> Dec 2001    115.18    101.74
#> Jan 2002    116.03    101.35
#> Feb 2002    116.52    101.66
#> Mar 2002    116.73    103.03
#> Apr 2002    117.13    102.03
#> May 2002    115.85    102.45
#> Jun 2002    115.15     99.10
#> Jul 2002    114.46     99.91
#> Aug 2002    116.86    102.49
#> Sep 2002    114.54    100.12
#> Oct 2002    113.29    101.34
#> Nov 2002    114.81    101.06
#> Dec 2002    112.02    100.55
#> Jan 2003    113.78    100.39
#> Feb 2003    114.20    101.44
#> Mar 2003    114.32    100.52
#> Apr 2003    113.91    101.79
#> May 2003    110.96    100.21
#> Jun 2003    110.75    100.87
#> Jul 2003    113.00    101.11
#> Aug 2003    112.00     98.55
#> Sep 2003    112.00    101.62
#> Oct 2003    113.93     99.70
#> Nov 2003    111.89     99.07
#> Dec 2003    112.61    100.22
#> Jan 2004    113.19    102.07
#> Feb 2004    115.13    100.30
#> Mar 2004    114.01    100.46
#> Apr 2004    113.99     99.58
#> May 2004    113.02     97.79
#> Jun 2004    114.69    100.45
#> Jul 2004    115.46    100.47
#> Aug 2004    111.58     98.05
#> Sep 2004    114.66     99.52
#> Oct 2004    115.49     99.17
#> Nov 2004    113.25    100.14
#> Dec 2004    113.45     98.41
#> Jan 2005    116.00     99.03
#> Feb 2005    113.64     98.21
#> Mar 2005    111.29     97.29
#> Apr 2005    114.63     97.85
#> May 2005    112.99     98.20
#> Jun 2005    112.89     97.69
#> Jul 2005    112.97     97.63
#> Aug 2005    112.41     99.65
#> Sep 2005    115.08     99.04
#> Oct 2005    111.91     98.29
#> Nov 2005    116.31    100.70
#> Dec 2005    114.87     99.18
#> Jan 2006    113.71     96.70
#> Feb 2006    113.10     97.68
#> Mar 2006    115.55     99.41
#> Apr 2006    113.82     98.55
#> May 2006    117.00     99.42
#> Jun 2006    117.12     98.44
#> Jul 2006    115.01    100.23
#> Aug 2006    115.52    100.48
#> Sep 2006    116.68     99.13
#> Oct 2006    115.67     99.56
#> Nov 2006    114.82     99.19
#> Dec 2006    117.07    100.05
#> Jan 2007    115.83    100.42
#> Feb 2007    116.74    100.87
#> Mar 2007    117.56    100.47
#> Apr 2007    116.15    102.49
#> May 2007    118.53    102.32
#> Jun 2007    117.65    101.16
#> Jul 2007    117.71     99.33
#> Aug 2007    118.95    102.21
#> Sep 2007    115.19     99.20
#> Oct 2007    117.69    102.73
#> Nov 2007    116.33    100.47
#> Dec 2007    116.30    101.62
#> Jan 2008    117.36    101.19
#> Feb 2008    118.75    100.96
#> Mar 2008    116.77    100.15
#> Apr 2008    118.20    100.62
#> May 2008    113.57     98.60
#> Jun 2008    113.04     97.14
#> Jul 2008    113.81     99.51
#> Aug 2008    111.82     98.29
#> Sep 2008    111.50     99.46
#> Oct 2008    109.09     99.21
#> Nov 2008    102.76     96.48
#> Dec 2008    101.00    100.86
#> Jan 2009     96.23     96.79
#> Feb 2009     95.10     97.72
#> Mar 2009     94.12     97.99
#> Apr 2009     93.85     97.51
#> May 2009     95.17     99.58
#> Jun 2009     96.82     99.51
#> Jul 2009     96.00     99.37
#> Aug 2009     96.77     97.86
#> Sep 2009     98.34     99.27
#> Oct 2009     97.26     98.02
#> Nov 2009     98.79     99.44
#> Dec 2009     97.38     97.60
#> Jan 2010     97.69     99.35
#> Feb 2010     97.27     99.27
#> Mar 2010     99.59    100.16
#> Apr 2010     99.62     98.94
#> May 2010    100.17    100.71
#> Jun 2010    100.05     99.45
#> Jul 2010    100.96    101.76
#> Aug 2010     99.02     99.46
#> Sep 2010    101.04    100.69
#> Oct 2010    100.56    100.43
#> Nov 2010    101.45    100.99
#> Dec 2010    102.58     98.79
#> Jan 2011    104.80    103.10
#> Feb 2011    105.35    102.64
#> Mar 2011    104.31    103.81
#> Apr 2011    102.98    102.12
#> May 2011    106.49    102.82
#> Jun 2011    102.32    104.54
#> Jul 2011    103.77    105.44
#> Aug 2011    102.45    102.10
#> Sep 2011    102.37    103.68
#> Oct 2011    102.38    104.07
#> Nov 2011    104.84    104.23
#> Dec 2011    103.58    101.81
#> Jan 2012    102.22    101.43
#> Feb 2012    100.87    102.89
#> Mar 2012    102.69    103.33
#> Apr 2012    101.18    103.21
#> May 2012    100.86    103.20
#> Jun 2012    100.73    103.31
#> Jul 2012    101.63    103.15
#> Aug 2012    103.32    103.49
#> Sep 2012    100.75    102.67
#> Oct 2012     99.27    104.24
#> Nov 2012     99.11    102.87
#> Dec 2012    100.10    104.88
#> Jan 2013     99.42    100.99
#> Feb 2013     99.60    101.53
#> Mar 2013     99.14    100.92
#> Apr 2013    101.32    102.00
#> May 2013    101.29    102.14
#> Jun 2013    100.33     96.73
#> Jul 2013    100.71    100.19
#> Aug 2013     99.36    101.25
#> Sep 2013    100.03    100.05
#> Oct 2013    100.25    100.38
#> Nov 2013    100.59    101.23
#> Dec 2013    100.21    101.79
#> Jan 2014     99.66    101.08
#> Feb 2014    100.64    102.10
#> Mar 2014    100.03    100.15
#> Apr 2014    100.58    102.62
#> May 2014     97.92    100.64
#> Jun 2014     99.66    100.69
#> Jul 2014    100.96    101.38
#> Aug 2014     98.72    100.20
#> Sep 2014    100.90    102.45
#> Oct 2014     99.56    100.74
#> Nov 2014     98.36     99.80
#> Dec 2014    101.59    101.80
#> Jan 2015    100.27    101.23
#> Feb 2015    100.38    100.05
#> Mar 2015    101.35    102.50
#> Apr 2015    100.89    101.99
#> May 2015    101.32    101.37
#> Jun 2015    102.27    103.65
#> Jul 2015    100.37    101.67
#> Aug 2015    102.97    102.29
#> Sep 2015    102.99    102.34
#> Oct 2015    102.70    102.32
#> Nov 2015    101.95    102.05
#> Dec 2015    102.54    101.20
#> Jan 2016    103.21    102.16
#> Feb 2016    102.00    102.46
#> Mar 2016    100.50    100.55
#> Apr 2016    102.66    100.85
#> May 2016    102.01     99.29
#> Jun 2016    101.07     99.02
#> Jul 2016    100.69     99.98
#> Aug 2016    103.40    101.02
#> Sep 2016    101.91    101.66
#> Oct 2016    100.84    100.03
#> Nov 2016    103.90    101.18
#> Dec 2016    102.61    101.07
#> Jan 2017    102.28     99.69
#> Feb 2017    101.44    100.72
#> Mar 2017    104.40    101.80
#> Apr 2017    103.01    100.48
#> May 2017    105.12    102.07
#> Jun 2017    104.18    101.03
#> Jul 2017    104.79    100.61
#> Aug 2017    104.59    101.25
#> Sep 2017    105.33    100.78
#> Oct 2017    107.95    101.05
#> Nov 2017    107.11    100.39
#> Dec 2017    107.44     99.15